pyproject.toml
[project]
name = "tarot-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai>=1.82.0",
    "pytest>=8.3.5",
    "python-telegram-bot>=22.1",
    "pyyaml>=6.0.2",
    "google-generativeai>=0.7.2",
    "Pillow>=10.4.0",
    "google-genai>=1.36.0",
]
